+++ scooter/scooter/cites-project/model/src/main/java/isel/sisinf/model/interfaces/ICard.java
@@ -3,12 +3,14 @@ package isel.sisinf.model.interfaces;
 import isel.sisinf.model.Client;
 import isel.sisinf.model.TypeOfCard;
 
+import java.math.BigDecimal;
+
 public interface ICard {
     public int getId();
     public void setId(int id);
 
-    public double getCredit();
-    public void setCredit(double credit);
+    public BigDecimal getCredit();
+    public void setCredit(BigDecimal credit);
 
     public TypeOfCard getTypeofcard();
     public void setTypeofcard(TypeOfCard typeofcard);
